refactor: dto 검증 로직 추가
