+++ src/main/java/roomescape/dto/request/AvailableTimeRequest.java
@@ -1,10 +1,12 @@
 package roomescape.dto.request;
 
+import com.fasterxml.jackson.annotation.JsonFormat;
 import jakarta.validation.constraints.FutureOrPresent;
 import jakarta.validation.constraints.NotNull;
 import java.time.LocalDate;
 
 public record AvailableTimeRequest(
+        @JsonFormat(pattern = "yyyy-mm-dd")
         @FutureOrPresent(message = "날짜는 현재보다 미래여야합니다.") LocalDate date,
         @NotNull(message = "테마 ID는 필수입니다.") Long themeId
 ) {
