+++ src/main/java/roomescape/dto/request/SignupRequest.java
@@ -1,11 +1,13 @@
 package roomescape.dto.request;
 
+import jakarta.validation.constraints.Email;
 import jakarta.validation.constraints.NotBlank;
-import roomescape.domain.Member;
-import roomescape.domain.Role;
+import roomescape.entity.Member;
+import roomescape.entity.Role;
 
 public record SignupRequest(
         @NotBlank(message = "이름은 비어있을 수 없습니다.") String name,
+        @Email
         @NotBlank(message = "이메일은 비어있을 수 없습니다.") String email,
         @NotBlank(message = "비밀번호는 비어있을 수 없습니다.") String password
 ) {
