+++ src/main/java/roomescape/dto/response/ReservationResponse.java
@@ -2,7 +2,7 @@ package roomescape.dto.response;
 
 import java.time.LocalDate;
 import java.time.LocalTime;
-import roomescape.domain.Reservation;
+import roomescape.entity.Reservation;
 
 public record ReservationResponse(Long id, String name, LocalTime time, LocalDate date, String themeName) {
     public static ReservationResponse from(Reservation reservation) {
