+++ src/main/java/roomescape/dto/response/ReservationTimeSlotResponse.java
@@ -1,7 +1,7 @@
 package roomescape.dto.response;
 
 import java.time.LocalTime;
-import roomescape.domain.ReservationSlot;
+import roomescape.entity.ReservationSlot;
 
 public record ReservationTimeSlotResponse(Long id, LocalTime time, boolean alreadyBooked) {
