+++ src/main/java/roomescape/dto/response/ThemeResponse.java
@@ -1,6 +1,6 @@
 package roomescape.dto.response;
 
-import roomescape.domain.Theme;
+import roomescape.entity.Theme;
 
 public record ThemeResponse(Long id, String description, String name, String thumbnail) {
